+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 12/02/2020
 - Add endpoint to get microservice details by id
+- Add status and url filters to get microservices endpoint
 
 # 11/02/2020
 - Merge control tower cron deployment process